requirements.txt
aiohttp == 3.7.4.post0
async-timeout == 3.0.1
attrs == 20.3.0
git+https://github.com/Pycord-Development/pycord
multidict == 5.1.0
typing-extensions == 3.7.4.3
yarl == 1.6.3
asyncpraw == 7.4.0
dbfn == 1.0.6
discord-components == 2.1.1
googletrans == 3.1.0a0
wikipedia == 1.4.0
googlesearch-python == 1.0.1
Toggle == 1.1.3
sympycalc == 1.0.1
google-api-python-client == 2.31.0
pymongo == 3.12.0